:art: DI 설정 체계화
폴더 구조 변경

관련 이슈 번호: #13
